fix(book): correção na criação de livros

+++ src/books/books.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { CreateBookDto } from './dto/create-book.dto';
 import { UpdateBookDto } from './dto/update-book.dto';
 import { Book } from './entities/book.entity';
@@ -9,6 +9,11 @@ export class BooksService {
   constructor(private readonly BooksRepository: BooksRepository) { }
 
   async create(createBookDto: CreateBookDto) {
+    const book = this.BooksRepository.findBookIsbn(createBookDto.isbn)
+
+    if (book) {
+      throw new HttpException('este livro ja esta cadastrado', HttpStatus.BAD_GATEWAY)
+    }
     return this.BooksRepository.createBook(createBookDto)
   }
 
@@ -17,18 +22,42 @@ export class BooksService {
   }
 
   searchBook(q: Book) {
+    if (!q.title && !q.author && !q.genre) {
+      throw new HttpException('para fazer a pesquisa precisa informa o titulo, autor ou gênero do livro',
+        HttpStatus.BAD_REQUEST)
+    }
+
     return this.BooksRepository.searchBook(q)
   }
 
   findOne(id: string) {
+    const book = this.BooksRepository.findOneBook(id)
+
+    if (!book) {
+      throw new HttpException('Este livro não esta cadastrado', HttpStatus.BAD_REQUEST)
+    }
+
     return this.BooksRepository.findOneBook(id)
   }
 
   update(id: string, updateBookDto: UpdateBookDto) {
+
+    const book = this.BooksRepository.findOneBook(id)
+
+    if (!book) {
+      throw new HttpException('Este livro não esta cadastrado', HttpStatus.BAD_REQUEST)
+    }
+
     return this.BooksRepository.updateBook(id, updateBookDto)
   }
 
   remove(id: string) {
+    const book = this.BooksRepository.findOneBook(id)
+
+    if (!book) {
+      throw new HttpException('Este livro não esta cadastrado', HttpStatus.BAD_REQUEST)
+    }
+
     return this.BooksRepository.removeBook(id)
   }
 }
